requirements.txt
Django==1.8
Pillow==3.1.1
argparse==1.2.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-filter==0.13.0
easy-thumbnails==2.3
psycopg2==2.6.1
sqlparse==0.1.19
wsgiref==0.1.2
